+++ Interface.cpp
@@ -75,7 +75,7 @@ Interface::Interface() {
 	//Get number of cores
 	SYSTEM_INFO sysinfo;
 	GetSystemInfo(&sysinfo);
-	numCPU = sysinfo.dwNumberOfProcessors;
+	numCPU = (size_t)sysinfo.dwNumberOfProcessors;
 
 	antiAliasing = true;
 	whiteBg = false;
